requirements.txt
ocean-spark-airflow-provider==0.1.3
PyMySQL
awscli
simple-salesforce
awswrangler
apache-airflow-providers-tableau==4.2.0
pyathena
great_expectations==0.16.13
airflow-provider-great-expectations
apache-airflow-providers-hashicorp
pandas==1.3.5
mysql-connector-python